requirements.txt
pyOpenSSL==22.1.0
requests==2.28.1
loguru==0.6.0